refactor(models): extract admin API request models into models/admin.py
Move SuspendRequest, TerminateRequest, ReviewRequest, ReasonRequest,
LogLevelBody (renamed from _LogLevelBody), UniverseConfigPatchRequest,
ActivatePromptRequest, and PreviewPromptRequest to the models layer.
Update routes/admin.py to import from tta.models.admin.

Reduces admin.py from 1,229 to 1,197 lines.
